refactor: move AMI to EC2
