+++ pkg/iac/adapters/cloudformation/aws/adapt.go
@@ -1,7 +1,6 @@
 package aws
 
 import (
-	"github.com/aquasecurity/trivy/pkg/iac/adapters/cloudformation/aws/ami"
 	"github.com/aquasecurity/trivy/pkg/iac/adapters/cloudformation/aws/apigateway"
 	"github.com/aquasecurity/trivy/pkg/iac/adapters/cloudformation/aws/athena"
 	"github.com/aquasecurity/trivy/pkg/iac/adapters/cloudformation/aws/cloudfront"
@@ -40,7 +39,6 @@ import (
 // Adapt adapts a Cloudformation AWS instance
 func Adapt(cfFile parser.FileContext) aws.AWS {
 	return aws.AWS{
-		AMI:           ami.Adapt(cfFile),
 		APIGateway:    apigateway.Adapt(cfFile),
 		Athena:        athena.Adapt(cfFile),
 		Cloudfront:    cloudfront.Adapt(cfFile),
